test: split admin experiment endpoint coverage
