+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -136,6 +136,8 @@ pub mod workflows;
 mod analytics_tests;
 #[cfg(all(test, feature = "admin"))]
 mod basic_endpoint_tests;
+#[cfg(all(test, feature = "admin-persist-sqlite"))]
+mod experiments_tests;
 #[cfg(feature = "admin")]
 mod handlers;
 #[cfg(all(test, feature = "admin"))]
